% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pca.R
\name{tidy_pca}
\alias{tidy_pca}
\title{PCA in a tidyverse framework}
\usage{
tidy_pca(data, ..., doCenter = TRUE, doScale = TRUE)
}
\arguments{
\item{data}{input data frame}

\item{...}{variables to include in PCA}

\item{doCenter}{passed to prcomp, center the data? TRUE / FALSE}

\item{doScale}{passed to prcomp, scale the data? TRUE / FALSE}
}
\value{
nested tibble with entries:\cr
  pca        = prcomp() output\cr
  data_aug   = original data with bound PC scores\cr
  pc_weights = tibble of PCA vector weights\cr
  pc_frac    = tibble of PCA eigenvalues and their cumulative (normalized) sum\cr
}
\description{
A simple wrapper for pca on a data frame
}
\examples{
df_pca <-
  iris \%>\%
  tidy_pca(
    Sepal.Length,
    Sepal.Width,
    Petal.Length,
    Petal.Width
  )

df_pca \%>\%
  unnest(pc_frac) \%>\%
  knitr::kable()

df_pca \%>\%
  unnest(data_aug) \%>\%
  ggplot(aes(PC1, PC2, color = Species)) +
  geom_point()
}
\seealso{
\code{\link{prcomp}} which this function wraps
}
\keyword{pca,}
\keyword{tibble}
